added dark theme ready to be merged in main
