+++ src/App.js
@@ -2,10 +2,10 @@ import React from 'react'
 import { HashRouter as Router, Routes, Route } from 'react-router-dom'
 import MyHome from './pages/MyHome'
 import AboutMe from './pages/AboutMe'
-import Header from './Header'
+import Header from './components/Header'
 import MySkills from './pages/MySkills'
 import ContactMe from './pages/ContactMe'
-import ThemeProviderWrapper from './ThemeProvider'
+import ThemeProviderWrapper from './theme/ThemeProvider'
 import { ThemeProvider, createTheme } from '@mui/material/styles'
 
 function App() {
