+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 import App from './App';
-import ThemeProviderWrapper from './ThemeProvider'
+import ThemeProviderWrapper from './theme/ThemeProvider'
 
 ReactDOM.render(
   <BrowserRouter>
